Moves InstantSearch inside AppHeader
